requirements.txt
flask
boto3
pika
Pillow
gunicorn